requirements.txt
astroid==2.15.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
greenlet==2.0.2
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
psycopg2==2.9.6
PyJWT==2.6.0
pylint==2.17.2
pylint-flask==0.6
pylint-plugin-utils==0.7
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.9
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
